Add QueryModification
