+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/queries/contract/HapiGetContractInfo.java
@@ -178,7 +178,7 @@ public class HapiGetContractInfo extends HapiQueryOp<HapiGetContractInfo> {
 
     @Override
     protected void submitWith(HapiSpec spec, Transaction payment) throws Throwable {
-        Query query = getContractInfoQuery(spec, payment, false);
+        Query query = maybeModified(getContractInfoQuery(spec, payment, false), spec);
         response = spec.clients().getScSvcStub(targetNodeFor(spec), useTls).getContractInfo(query);
         ContractInfo contractInfo = response.getContractGetInfo().getContractInfo();
         if (verboseLoggingOn) {
