+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/queries/token/HapiGetTokenInfo.java
@@ -443,7 +443,7 @@ public class HapiGetTokenInfo extends HapiQueryOp<HapiGetTokenInfo> {
 
     @Override
     protected void submitWith(HapiSpec spec, Transaction payment) {
-        Query query = getTokenInfoQuery(spec, payment, false);
+        Query query = maybeModified(getTokenInfoQuery(spec, payment, false), spec);
         response = spec.clients().getTokenSvcStub(targetNodeFor(spec), useTls).getTokenInfo(query);
         if (verboseLoggingOn) {
             LOG.info("Info for '{}': {}", () -> token, response.getTokenGetInfo()::getTokenInfo);
